requirements.txt
nba_api>=1.5.2
dash>=2.17.0
dash-bootstrap-components>=1.6.0
plotly>=5.22.0
pandas>=2.2.2
numpy>=1.26.0
requests>=2.32.3
python-dotenv>=1.0.1
diskcache>=5.6.3
